feat: refactoring for folders
