+++ constants/cipherConstants.js
