+++ constants/cliConstants.js
