+++ modules/ReadableForCipherSource.js
