+++ modules/WritableForCipher.js
