+++ my_ciphering_cli.js
@@ -1,15 +1,15 @@
 
 const { pipeline } = require('stream');
 
-const { ReadableForCipherSource } = require('./ReadableForCipherSource');
-const { WritableForCipher } = require('./WritableForCipher');
+const { ReadableForCipherSource } = require('./modules/ReadableForCipherSource');
+const { WritableForCipher } = require('./modules/WritableForCipher');
 const {
     getArrayOfTransformStreams,
     getChainConfig,
     getInputFile,
     getOutputFile,
     showError,
-} = require('./utils');
+} = require('./utils/utils');
 
 const args = process.argv.slice(2);
